requirements.txt
streamlit
google-cloud-bigquery
pycountry
wordcloud
db-dtypes
